Rename MS-sculpt-mobile to MS_scuplpt_mobile
Subprojects with "-" have problems, since the build system creates
a define containing the subproject name, and "-" is illegal in
macro names.
