+++ keyboards/handwired/MS_sculpt_mobile/MS_sculpt_mobile.c
@@ -1,1 +1,1 @@
-#include "MS-sculpt-mobile.h"
+#include "MS_sculpt_mobile.h"
